Updated to f16

+++ utils.py
@@ -76,7 +76,7 @@ def load_data(path=None):
     print('Searching at: {}'.format(search_path))
     paths = glob(search_path)
     fpath = 'memmap_numpy_images.dat'
-    data = np.memmap(fpath, dtype='float32', mode='w+',
+    data = np.memmap(fpath, dtype='float16', mode='w+',
                      shape=(len(paths), 64, 64, 3))
     pool = multiprocessing.Pool()
     counter = 0
